pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "acrossfc-api"
description = "Across FC Automation and Backend API"
version = "1.0"
authors = [
    {name = "Kohen Chia", email = "kohenchia@gmail.com"}
]
requires-python = ">= 3.12"
readme = "README.md"
dependencies = [
    'click',
    'requests',
    'aiohttp',
    'gql',
    'tabulate',
    'peewee',
    'google-api-python-client',
    'google-auth-httplib2',
    'google-auth-oauthlib',
    'urllib3==1.26.17'
]

[project.optional-dependencies]
dev = [
    "awscli",
    "boto3"
]

[project.scripts]
etl = "acrossfc.etl:etl"
axr = "acrossfc.analytics:axr"
axs = "acrossfc.api.submissions:axs"

[tool.setuptools.packages.find]
include = ["acrossfc"]
